switch to strategy pattern
