+++ src/bunit.web.query/Labels/Strategies/ILabelTextQueryStrategy.cs
@@ -1,0 +1,8 @@
+﻿using AngleSharp.Dom;
+
+namespace Bunit;
+
+internal interface ILabelTextQueryStrategy
+{
+	IElement? FindElement(IRenderedFragment renderedFragment, string labelText);
+}
